% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify_prs_into_ntiles.R
\name{decile_breaks}
\alias{decile_breaks}
\title{Classify Subjects Into Intervals According To PRS Values}
\usage{
decile_breaks(prs_values)
}
\arguments{
\item{prs_values}{A numeric vector with PRS values.}
}
\value{
A character vector with intervals in which each PRS value was classified.
}
\description{
Classification of individuals into intervals according to deciles and to breaks present in Mavaddat et al., 2019.
}
\examples{
#TBD
}
