% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_control_stats.R
\name{get_control_stats}
\alias{get_control_stats}
\title{Controls' PRS Mean and Standard Deviation}
\usage{
get_control_stats(df)
}
\arguments{
\item{df}{A dataframe with columns status, and "prs" in the start of all PRS columns.}
}
\value{
A tibble with columns named "prs\[something\]_mean", "prs\[something\]_sd" for all PRS columns, where mean and sd were calculated based on controls only (status == 0).
}
\description{
Obtain mean and standard deviation for only controls of all PRSss
}
\examples{
#TBD
}
