% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models_per_sd.R
\name{model_with_prs}
\alias{model_with_prs}
\title{Models per PRS Standard Deviation}
\usage{
model_with_prs(
  train_ctrl_stats,
  test_ctrl_stats,
  log_reg,
  norm_prs,
  recipe_var = NULL
)
}
\arguments{
\item{train_ctrl_stats}{Training set of the PRS data.}

\item{test_ctrl_stats}{Testing set of the PRS data.}

\item{log_reg}{A parsnip logistic regression model.}

\item{norm_prs}{Character vector with the name of the normalized PRS column.}

\item{recipe_var}{A recipe from recipes::recipe() object, to be provided optionally, in case the recipe originally included in the model (status ~ prs + scaled_centered_pc(1:10) + age_analysis) is not what you need for your PRS.}
}
\value{
A list with OR, AUC, delta AUC, and ROC curve for AUC with and without PRS.
}
\description{
Functions to obtain models with PRS + covariates, with only covariates or only PRS
}
\examples{
# TBD
}
