% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics_per_interval.R
\name{per_interval_metrics}
\alias{per_interval_metrics}
\title{Metrics per Interval of PRS Values}
\usage{
per_interval_metrics(dataset, prs_col, intervals_mode)
}
\arguments{
\item{dataset}{A dataframe with columns status, age_analysis (age at analysis for controls or age at diagnosis for cases), and impersonate_code (individuals id).}

\item{prs_col}{Name of PRS column (character).}

\item{intervals_mode}{Either "decile" or "mavaddat" to define the intervals that should be created in the dataset. See classify_decile() and classify_mavaddat() to understand the difference.}
}
\value{
A list with OR, AUC, delta AUC, and ROC curve for AUC with and without PRS.
}
\description{
Get OR metrics for a certain PRS per defined intervals according to PRS values.
}
\examples{
#TBD
}
