% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/per_sd_metrics.R
\name{per_sd_metrics}
\alias{per_sd_metrics}
\title{Metrics per Standard Deviation}
\usage{
per_sd_metrics(dataset, prs_col, seed, recipe_var = NULL)
}
\arguments{
\item{dataset}{A dataframe with columns status, age at analysis or diagnosis, impersonate code and id.}

\item{prs_col}{Name of PRS column (character).}

\item{seed}{A random number to be set as a seed for the training and testing sampling to be reproducible.}

\item{recipe_var}{A recipe from recipes::recipe() object, to be provided optionally, in case the recipe originally included in the model (status ~ prs + scaled_centered_pc(1:10) + age_analysis) is not what you need for your PRS.}
}
\value{
A list with OR, AUC, delta AUC, and ROC curve for AUC with and without PRS.
}
\description{
Get OR and AUC metrics for a certain PRS
}
\examples{
# PCs 1 to 10
n_cols <- 10
n_rows <- 100
data_matrix <- replicate(
  n_cols,
  rnorm(n_rows, mean = runif(1, 0, 10), sd = runif(1, 1, 5))
)
df_pcs <- as.data.frame(data_matrix)
colnames(df_pcs) <- paste0("pc", 1:n_cols)
# Data mock
data_mock <- data.frame(
  impersonate_code = paste0(
    sample(LETTERS, 100, replace = TRUE),
    sample(LETTERS, 100, replace = TRUE),
    sample(LETTERS, 100, replace = TRUE),
    sample(1:9, 100, replace = TRUE),
    sample(1:9, 100, replace = TRUE),
    sample(1:9, 100, replace = TRUE)
  ),
  status = as.factor(sample(c(0, 1), size = 100, replace = TRUE)),
  age_analysis = round(runif(100, 19, 80)),
  tier1 = as.factor(sample(
    c(0, 1),
    size = 100,
    prob = c(0.8, 0.2),
    replace = TRUE
  )),
  prs_test = rnorm(n = 100),
  version = sample(c("v1", "v2"), size = 100, replace = TRUE)
) |>
  cbind(df_pcs)

# Arguments
seed <- 28
prs_col_mock <- "prs_test"

per_sd_metrics(data_mock, prs_col_mock, seed)

}
