% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_prs_or_auc.R
\name{get_control_stats}
\alias{get_control_stats}
\title{Get OR and AUC metrics for a certain PRS}
\usage{
get_control_stats(df)
}
\arguments{
\item{dataset}{A dataframe with columns status, age, tier1(0/1 for presence of pathogenic or likely-pathogenic variants in tier1 list for the disease evaluated).}

\item{prs_col}{Name of PRS column (character).}

\item{seed}{A random number to be set as a seed for the training and testing sampling to be reproducible.}
}
\value{
A list with OR, AUC, delta AUC, and ROC curve for AUC with and without PRS.
}
\description{
Get OR and AUC metrics for a certain PRS
}
\examples{
# get_prs_or_auc(breast_v2_v3_mut, x, seed = 28)
# TBD
}
